refactor: 特殊任务仅允许放在主任务前后
