+++ src/i18n/locales/zh-CN.ts
@@ -207,6 +207,7 @@ export default {
     taskSkippedController: '任务 "{{taskName}}" 不支持当前控制器',
     taskSkippedResource: '任务 "{{taskName}}" 不支持当前资源',
     noCompatibleTasks: '没有兼容当前控制器和资源的任务',
+    invalidTaskOrder: '特殊任务只能放在普通任务之前或之后，不能夹在普通任务中间',
     // 自动连接相关
     autoConnect: {
       searching: '搜索设备...',
