+++ src/utils/taskSegmentation.ts
@@ -10,7 +10,7 @@ export function shouldSkipScreenshot(taskName: string): boolean {
   return shouldSkipMxuScreenshot(taskName) || isPretaskName(taskName);
 }
 
-/** 三段式任务切分结果：前置特殊 / 中间游戏 / 收尾特殊 */
+/** 三段式任务切分结果：连接前特殊 / 连接中普通 / 连接后特殊 */
 export interface ThreeSegmentSplit<T> {
   leading: T[];
   middle: T[];
@@ -18,10 +18,12 @@ export interface ThreeSegmentSplit<T> {
 }
 
 /**
- * 将任务列表切成最多三段：
- * - leading：队首连续 MXU 特殊任务
- * - middle：第一个游戏任务到最后一个游戏任务（含夹在游戏之间的特殊任务）
- * - trailing：队尾连续 MXU 特殊任务
+ * 将任务列表按「连接前 / 连接中 / 连接后」严格三段切分：
+ * - leading：第一个普通任务之前的特殊任务（连接前，Dummy 执行）
+ * - middle：所有普通任务（连接中，真机执行）
+ * - trailing：第一个普通任务及之后出现的全部特殊任务，含夹心与队尾（连接后，Dummy 执行）
+ *
+ * 注意：入参应为已排除 pretask 的可执行任务，`shouldSkipScreenshot` 在此等价于「特殊任务」。
  */
 export function splitTasksIntoThreeSegments<T extends { taskName: string }>(
   tasks: T[],
@@ -30,56 +32,42 @@ export function splitTasksIntoThreeSegments<T extends { taskName: string }>(
     return { leading: [], middle: [], trailing: [] };
   }
 
-  let leadingEnd = 0;
-  while (leadingEnd < tasks.length && shouldSkipScreenshot(tasks[leadingEnd].taskName)) {
-    leadingEnd += 1;
+  const firstNormalIdx = tasks.findIndex((t) => !shouldSkipScreenshot(t.taskName));
+
+  // 没有普通任务：全部作为连接前特殊任务（全程 Dummy）
+  if (firstNormalIdx === -1) {
+    return { leading: [...tasks], middle: [], trailing: [] };
   }
 
-  let trailingStart = tasks.length;
-  while (trailingStart > leadingEnd && shouldSkipScreenshot(tasks[trailingStart - 1].taskName)) {
-    trailingStart -= 1;
-  }
+  const leading = tasks.slice(0, firstNormalIdx);
+  const rest = tasks.slice(firstNormalIdx);
+  const middle = rest.filter((t) => !shouldSkipScreenshot(t.taskName));
+  const trailing = rest.filter((t) => shouldSkipScreenshot(t.taskName));
 
-  return {
-    leading: tasks.slice(0, leadingEnd),
-    middle: tasks.slice(leadingEnd, trailingStart),
-    trailing: tasks.slice(trailingStart),
-  };
-}
-
-/** 是否存在需要 Dummy 空 controller 的特殊段（前置或收尾） */
-export function hasSpecialSegments(segments: ThreeSegmentSplit<unknown>): boolean {
-  return segments.leading.length > 0 || segments.trailing.length > 0;
-}
-
-/** 是否应走分段运行（存在特殊段且与游戏段组合） */
-export function needsSegmentedRun(segments: ThreeSegmentSplit<unknown>): boolean {
-  return hasSpecialSegments(segments);
-}
-
-export type MxuTaskSegmentKind = 'leading' | 'middle' | 'trailing';
-
-export interface MxuTaskExecutionSegment<T> {
-  kind: MxuTaskSegmentKind;
-  tasks: T[];
-  useDummyController: boolean;
+  return { leading, middle, trailing };
 }
 
 /**
- * 将任务列表分组为可执行批次：
- * - leading：队首特殊任务，使用 Dummy Controller
- * - middle：中间游戏段，使用当前游戏控制器
- * - trailing：队尾特殊任务，使用 Dummy Controller
+ * 校验任务顺序是否满足三段式不变式 `S* N* S*`：
+ * 忽略 pretask 后，特殊任务只能位于普通任务块之前或之后，不得夹在两个普通任务之间。
+ * 用于拖拽/移动排序时拒绝会破坏分段的操作。
  */
-export function getMxuTaskExecutionSegments<T extends { taskName: string }>(
-  tasks: T[],
-): MxuTaskExecutionSegment<T>[] {
-  const { leading, middle, trailing } = splitTasksIntoThreeSegments(tasks);
-  return [
-    { kind: 'leading' as const, tasks: leading, useDummyController: true },
-    { kind: 'middle' as const, tasks: middle, useDummyController: false },
-    { kind: 'trailing' as const, tasks: trailing, useDummyController: true },
-  ].filter((segment) => segment.tasks.length > 0);
+export function isValidTaskOrder<T extends { taskName: string }>(tasks: T[]): boolean {
+  // pretask 处于独立区块，不参与主列表的特殊/普通分段校验
+  const relevant = tasks.filter((t) => !isPretaskName(t.taskName));
+
+  // phase: 0=连接前特殊, 1=普通, 2=连接后特殊
+  let phase: 0 | 1 | 2 = 0;
+  for (const task of relevant) {
+    const isSpecial = shouldSkipScreenshot(task.taskName);
+    if (isSpecial) {
+      if (phase === 1) phase = 2;
+    } else {
+      if (phase === 0) phase = 1;
+      else if (phase === 2) return false; // 普通任务出现在连接后特殊任务之后，非法
+    }
+  }
+  return true;
 }
 
 export type { SelectedTask };
